finishing (for now) touches on apu and "getting 60fps"
timerBeginPeriod did increase sleep precision BUT things still seem to
run a little fast sometimes. i don't think this is because of using
4190000 in place of 4194304 in a few places, that would be an ~0.1%
error where what i see is closer to 6% (getting 64fps instead of 60).

the timing loop should account for extra sleep, so i'm at a loss for the
extra 4fps. in the mean time i've synced the main loop's sleep to the
depth of the audio queue, so that if there are more than 10 samples in
the queue (>5.8ms of audio), the main loop will start sleeping
increasingly longer to maintain an audio depth of 10-or-fewer samples.

also i think audio channels are not stopped properly. should channels
stop when volume hits 15 and overflows? or am i handling length wrong...