hdma transfers mask out low four bits
Oracle of Seasons and Oracle of Ages do hdma transfers to destinations
with the low bit set. without this change, data would be transferred
forward by one byte and result in... many graphical issues. would love
to know what ends up with the +1 addresses, and if that's an emulator
bug, game bug, or neither..