branch and call tracing
this is useful really only to try figuring out if call/interrupts are
being handled.... correctly.

specifically, if an interrupt arrives and `reti` does not return control
to the address an interrupt was received at, that's *probably* a bug. if
a `call` happens and `ret` does not return to the caller, that can be
weird too. though some games do tail calls to other procedures and make
this heuristic *quite* incorrect.