do not run the lcd when lcd is off
aside from simply drawing half-loaded scenes and being *wrong*, this
tickles bugs in released games.

super mario world 2: 6 golden coins tries to be interrupt-safe while
loading data from rom for the save select scene, right after pressing
start after starting the game. this copying happens in the routine at
0x2b8e - `a24e` is an address for the rom bank that code wants to be
executing in, which interrupt handlers restore as the true rom bank
before executing `iret`. each store to `2100` then actually changes the
rom bank.

this ordering is important! if a rom bank is indicated in a24e, then an
interrupt occurs, the interrupt handler will restore the desired rom
bank which a subsequent store to 2100 will also apply. if an interrupt
happens after the rom stored to 2100, the interrupt handler will restore
the desired rom bank and all wll continue like normal.

this is a consistent pattern for all sequences in the routine at 2b8e
... except for the very last one at 0x2bd6. unlike other instances, this
sets a rom bank of 0xc, but does not store that preference to 0xa24e. so
if an interrupt happens at 2bdb or later, the interrupt handler will
incorrectly restore the rom bank to 0x1b and a return from this
routine will actually return to wildly incorrect code. bug.

oracle of ages executes incorrectly for similar reasons if a vblank
arrives while preparing the scene transition after nayru becomes
possessed early in the game, just after travelling to the past.

these were both fixed by... turning off the lcd when it is off.