swap from Vec<Pixel> to [Pixel; 160]
even though the vec would be grown to 160 items and never resized or
grown larger, using Vec means insertion still has to bounds check for a
condition that is always true, and include code for growing a vec that
will never grow.

using a fixed-size array of the right length results in less (and
better!) code, slightly better throughput.