do not test for or run dma transfers in the hot path
proably should factor this out into some kind of utility function, but
the primary purpose of this change is to move dma transfer checking and
execution out of `GBC::run`. in that function, we regularly check for
dma transfers that almost never happen, and grow the stack frame size
(and some of the subsequent initialization).. again, for code that is
very unlikely to run.

this was some (~2-5%?) improvement on windows and linux