factor gbc state out into a discrete struct
also remove some straggler eprintln that have little value other than
growing hot function sizes with unused logging code

this is primarily interesting to:
* shrink the size of MemoryMapping
* initialize less when constructing MemoryMapping before calling `Cpu::step`.

MemoryMapping was ~82 bytes before, including a convoluted branchless
simd sequence to pick which of the boot rom or cartridge were the active
cartridge to emulate. after this, MemoryMapping is ~25 bytes, and the
per-instruction simd shuffle is no longer present.

this was another 7%ish improvement.

additionally, stop setting `management_bits[LY]`. when reading LY, fetch
it directly from `lcd.ly`. this is another pointless per-frame update,
and was a 3-5% improvement.